requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
bip-utils==2.9.3
cbor2==5.6.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coincurve==20.0.0
construct==2.10.70
crcmod==1.7
crypto_cpp_py==1.4.4
cryptography==43.0.1
cytoolz==0.12.3
ecdsa==0.18.0
ed25519-blake2b==1.4.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-typing==5.0.0
eth-utils==5.0.0
frozenlist==1.4.1
hexbytes==1.2.1
hidapi==0.14.0.post2
idna==3.10
intelhex==2.3.0
lark==1.2.2
ledgerwallet==0.5.1
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow_dataclass==8.7.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pillow==10.4.0
poseidon_py==0.1.5
protobuf==3.20.3
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
PyNaCl==1.5.0
python-decouple==3.8
requests==2.32.3
setuptools==75.1.0
six==1.16.0
starknet-py==0.24.2
sympy==1.11.1
tabulate==0.9.0
toml==0.10.2
toolz==0.12.1
typeguard==4.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
